Move layouts to core
