+++ docs/video/layout-demo.tape
@@ -1,0 +1,21 @@
+Source shared_.tape
+
+# Setup
+Hide
+Type "jbang layout-demo"
+Enter
+Sleep 2
+Show
+
+# Recording
+Sleep 2s
+Type "2"
+Sleep 2s
+Type "3"
+Sleep 2s
+Type "+"
+Sleep 1s
+Type "4"
+Sleep 2s
+Type "1"
+Sleep 2s
