+++ tamboui-core/demos/flow-demo/build.gradle.kts
@@ -1,0 +1,14 @@
+plugins {
+    id("dev.tamboui.demo-project")
+}
+
+description = "Demo showcasing the Flow layout with wrap layout"
+
+demo {
+    displayName = "Flow Layout"
+    tags = setOf("flow", "layout", "wrap")
+}
+
+application {
+    mainClass.set("dev.tamboui.demo.FlowDemo")
+}
