+++ tamboui-toolkit/src/main/java/dev/tamboui/toolkit/elements/ColumnOrderConverter.java
@@ -5,7 +5,7 @@
 package dev.tamboui.toolkit.elements;
 
 import dev.tamboui.style.PropertyConverter;
-import dev.tamboui.widgets.columns.ColumnOrder;
+import dev.tamboui.layout.columns.ColumnOrder;
 
 import java.util.Optional;
