+++ tamboui-toolkit/src/test/java/dev/tamboui/toolkit/elements/ColumnsElementTest.java
@@ -14,7 +14,7 @@ import dev.tamboui.style.Style;
 import dev.tamboui.terminal.Frame;
 import dev.tamboui.toolkit.element.DefaultRenderContext;
 import dev.tamboui.toolkit.element.RenderContext;
-import dev.tamboui.widgets.columns.ColumnOrder;
+import dev.tamboui.layout.columns.ColumnOrder;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
