--- tamboui-widgets/demos/flow-demo/build.gradle.kts
@@ -1,13 +1,0 @@
-plugins {
-    id("dev.tamboui.demo-project")
-}
-
-description = "Demo showcasing the Flow widget with wrap layout"
-
-demo {
-    tags = setOf("flow", "layout", "wrap")
-}
-
-application {
-    mainClass.set("dev.tamboui.demo.FlowDemo")
-}
