--- tamboui-widgets/demos/grid-demo/build.gradle.kts
@@ -1,13 +1,0 @@
-plugins {
-    id("dev.tamboui.demo-project")
-}
-
-description = "Demo showcasing Grid widget with CSS Grid-inspired layout"
-
-demo {
-    tags = setOf("grid", "layout", "block", "paragraph")
-}
-
-application {
-    mainClass.set("dev.tamboui.demo.GridDemo")
-}
