--- tamboui-widgets/demos/stack-demo/build.gradle.kts
@@ -1,13 +1,0 @@
-plugins {
-    id("dev.tamboui.demo-project")
-}
-
-description = "Demo showcasing the Stack widget with overlapping layers"
-
-demo {
-    tags = setOf("stack", "layout", "overlay")
-}
-
-application {
-    mainClass.set("dev.tamboui.demo.StackDemo")
-}
